refactor(models): keep invoke contract within lint budgets
