+++ pkg/transports/cli/climanifestcobra/models_constructor_test.go
@@ -381,14 +381,27 @@ func TestModelsCommandRegistersPositionalsAndFlagsFromManifest(t *testing.T) {
 	if invoke.Use != "invoke <model-name>" {
 		t.Fatalf("invoke use = %q", invoke.Use)
 	}
+	assertModelsInvokeFlagsRegistered(t, invoke)
+	if err := invoke.ParseFlags([]string{"--operation", "TTS", "--offline", "--input", "audio=@meeting.wav", "--input", "prompt=hint", "--text", "hello", "--output", "speech.wav"}); err != nil {
+		t.Fatal(err)
+	}
+	assertModelsInvokeParsedFlags(t, invoke)
+	if err := invoke.ParseFlags([]string{"--operation", "INVALID"}); err == nil {
+		t.Fatal("invalid manifest operation choice was accepted")
+	}
+}
+
+func assertModelsInvokeFlagsRegistered(t *testing.T, invoke *cobra.Command) {
+	t.Helper()
 	for _, name := range []string{"operation", "offline", "input", "text", "output", "output-map", "port"} {
 		if invoke.Flags().Lookup(name) == nil {
 			t.Fatalf("manifest flag %q was not registered", name)
 		}
 	}
-	if err := invoke.ParseFlags([]string{"--operation", "TTS", "--offline", "--input", "audio=@meeting.wav", "--input", "prompt=hint", "--text", "hello", "--output", "speech.wav"}); err != nil {
-		t.Fatal(err)
-	}
+}
+
+func assertModelsInvokeParsedFlags(t *testing.T, invoke *cobra.Command) {
+	t.Helper()
 	for name, want := range map[string]string{"operation": "TTS", "text": "hello"} {
 		got, getErr := invoke.Flags().GetString(name)
 		if getErr != nil || got != want {
@@ -407,9 +420,6 @@ func TestModelsCommandRegistersPositionalsAndFlagsFromManifest(t *testing.T) {
 	if offlineErr != nil || !offline {
 		t.Fatalf("offline flag = %t, %v; want true", offline, offlineErr)
 	}
-	if err := invoke.ParseFlags([]string{"--operation", "INVALID"}); err == nil {
-		t.Fatal("invalid manifest operation choice was accepted")
-	}
 }
 
 func TestModelsInvokeManifestDeclaresOneOfflineFlagAndBinding(t *testing.T) {
@@ -422,10 +432,25 @@ func TestModelsInvokeManifestDeclaresOneOfflineFlagAndBinding(t *testing.T) {
 		t.Fatal(err)
 	}
 
+	assertModelsInvokeOfflineContract(t, invoke)
+}
+
+func assertModelsInvokeOfflineContract(t *testing.T, invoke climanifest.Command) {
+	t.Helper()
 	offline, ok := invoke.FlagByLong("offline")
 	if !ok {
 		t.Fatal("models invoke manifest is missing the offline flag")
 	}
+	assertModelsInvokeOfflineFlag(t, offline)
+	assertModelsInvokeHasOneVisibleOfflineFlag(t, invoke)
+	binding, ok := invoke.HandlerBindings["you.models.invoke.binding.offline"]
+	if !ok || binding.InputID != offline.ID {
+		t.Fatalf("offline handler binding = %#v; want input %q", binding, offline.ID)
+	}
+}
+
+func assertModelsInvokeOfflineFlag(t *testing.T, offline climanifest.Flag) {
+	t.Helper()
 	if offline.ID != "you.models.invoke.flag.offline" || offline.Scope != "local" || offline.ValueType != "bool" || offline.Required || offline.Repeatable || offline.Visibility != "visible" {
 		t.Fatalf("offline flag = %#v; want one visible local optional bool flag", offline)
 	}
@@ -438,7 +463,10 @@ func TestModelsInvokeManifestDeclaresOneOfflineFlagAndBinding(t *testing.T) {
 	if offline.NoOptionValue == nil || offline.NoOptionValue.Boolean == nil || !*offline.NoOptionValue.Boolean {
 		t.Fatalf("offline no-option value = %#v; want typed true", offline.NoOptionValue)
 	}
+}
 
+func assertModelsInvokeHasOneVisibleOfflineFlag(t *testing.T, invoke climanifest.Command) {
+	t.Helper()
 	visibleOfflineFlags := 0
 	for _, flag := range invoke.Flags {
 		if flag.Visibility == "visible" && flag.Long == "offline" {
@@ -447,11 +475,6 @@ func TestModelsInvokeManifestDeclaresOneOfflineFlagAndBinding(t *testing.T) {
 	}
 	if visibleOfflineFlags != 1 {
 		t.Fatalf("visible offline flags = %d; want exactly one", visibleOfflineFlags)
-	}
-
-	binding, ok := invoke.HandlerBindings["you.models.invoke.binding.offline"]
-	if !ok || binding.InputID != offline.ID {
-		t.Fatalf("offline handler binding = %#v; want input %q", binding, offline.ID)
 	}
 }
